Reorganized files so that things that were used globally were in their own separate directories at the top rather than being parents of the things that used them.
