+++ src/_Bot.ts
@@ -2,8 +2,8 @@ import { Client, GatewayIntentBits, Partials } from "discord.js";
 import { DISCORD_TOKEN, MONGODB_SRV } from "./secret";
 
 //import event listeners
-import commandListener from "./interaction/util.InteractionListener";
-import onlineLogger from "./util.ReadyListener";
+import commandListener from "./listeners/util.InteractionListener";
+import onlineLogger from "./listeners/util.ReadyListener";
 
 // TODO: converting this to an import disalows the useNewUrlParser option for some reason. When done with refactoring figure out if converting breaks it or not
 //import database connection thing
