+++ src/command/interface.Command.ts
