+++ src/pieces/courseManagement/Command.RemoveCourse.ts
@@ -1,6 +1,6 @@
 import { CommandInteraction, Client, ApplicationCommandOptionType, CategoryChannel } from 'discord.js';
-import { Course, courseModel } from '../../../generalModels/Course';
-import { Command } from '../interface.Command';
+import { Course, courseModel } from '../../generalModels/Course';
+import { Command } from '../../command/interface.Command';
 
 
 export const removeCourse: Command = {
